+++ elastic-job-lite-executor/elastic-job-lite-dataflow-executor/src/main/resources/META-INF/services/org.apache.shardingsphere.elasticjob.lite.executor.type.JobItemExecutor
@@ -15,4 +15,4 @@
 # limitations under the License.
 #
 
-org.apache.shardingsphere.elasticjob.lite.executor.type.impl.ScriptJobExecutor
+org.apache.shardingsphere.elasticjob.lite.dataflow.executor.DataflowJobExecutor
