+++ elastic-job-lite-executor/elastic-job-lite-dataflow-executor/src/test/java/org/apache/shardingsphere/elasticjob/lite/dataflow/job/BatchDataflowJob.java
@@ -7,7 +7,7 @@
  * the License.  You may obtain a copy of the License at
  *
  *     http://www.apache.org/licenses/LICENSE-2.0
- *
+ *  
  * Unless required by applicable law or agreed to in writing, software
  * distributed under the License is distributed on an "AS IS" BASIS,
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
@@ -15,10 +15,9 @@
  * limitations under the License.
  */
 
-package org.apache.shardingsphere.elasticjob.lite.integrate.fixture.dataflow;
+package org.apache.shardingsphere.elasticjob.lite.dataflow.job;
 
 import org.apache.shardingsphere.elasticjob.lite.api.job.ShardingContext;
-import org.apache.shardingsphere.elasticjob.lite.api.job.type.DataflowJob;
 
 import java.util.Arrays;
 import java.util.List;
