+++ elastic-job-lite-lifecycle/pom.xml
@@ -37,6 +37,11 @@
             <artifactId>elastic-job-lite-simple-executor</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere.elasticjob</groupId>
+            <artifactId>elastic-job-lite-dataflow-executor</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
         
         <dependency>
             <groupId>org.projectlombok</groupId>
