+++ elastic-job-lite-spring/pom.xml
@@ -36,6 +36,11 @@
             <artifactId>elastic-job-lite-simple-executor</artifactId>
             <version>${project.parent.version}</version>
         </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere.elasticjob</groupId>
+            <artifactId>elastic-job-lite-dataflow-executor</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
         
         <dependency>
             <groupId>org.springframework</groupId>
