+++ examples/elastic-job-example-jobs/pom.xml
@@ -35,6 +35,11 @@
             <groupId>org.apache.shardingsphere.elasticjob</groupId>
             <artifactId>elastic-job-lite-simple-executor</artifactId>
         </dependency>
+        <dependency>
+            <groupId>org.apache.shardingsphere.elasticjob</groupId>
+            <artifactId>elastic-job-lite-dataflow-executor</artifactId>
+            <version>${project.parent.version}</version>
+        </dependency>
         
         <dependency>
             <groupId>org.springframework</groupId>
